Move version parameter to web.xml

+++ src/main/java/com/ylysenkova/movieland/web/response/MovieWithReviewResponse.java
@@ -5,6 +5,7 @@ import com.ylysenkova.movieland.model.Genre;
 import com.ylysenkova.movieland.model.Movie;
 import com.ylysenkova.movieland.model.Review;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class MovieWithReviewResponse {
@@ -15,7 +16,7 @@ public class MovieWithReviewResponse {
     private int yearOfRelease;
     private String description;
     private double rating;
-    private double price;
+    private BigDecimal price;
     private String picturePath;
     private List<Country> countries;
     private List<Genre> genres;
@@ -89,11 +90,11 @@ public class MovieWithReviewResponse {
         this.rating = rating;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
